requirements.txt
aniso8601==9.0.1
ansible==6.5.0
ansible-core==2.13.5
ansible-vault-win==1.0.0
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cleo==0.8.1
click==8.0.4
clikit==0.6.2
colorama==0.4.4
coverage==6.4.1
crashtest==0.3.1
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.5
Django==4.0.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
environ==1.0
factory-boy==3.2.1
Faker==14.2.0
filelock==3.8.0
Flask==2.0.3
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.6.0
keyring==23.8.2
lockfile==0.12.2
MarkupSafe==2.1.1
marshmallow==3.16.0
msgpack==1.0.4
mypy==0.961
mypy-extensions==0.4.3
numpy==1.23.3
oauthlib==3.2.1
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.1.14
poetry-core==1.0.8
prettytable==3.2.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pylev==1.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resolvelib==0.8.1
scipy==1.9.1
shellingham==1.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
SQLAlchemy==1.4.32
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
tzdata==2022.4
urllib3==1.26.11
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
zipp==3.8.1
